requirements/base.txt
# Django
Django==1.9.1

# Database
dj-database-url==0.3.0
psycopg2==2.6.1
SQLAlchemy==1.0.11

# Sysadmin / Ops
gunicorn==19.4.5
django-configurations==1.0

# Static files
whitenoise==2.0.6

# Address / Localisation
django-localflavor==1.2

# Models & Fields
django-model-utils==2.4
django-phonenumber-field==1.0.0

# Users and Authentication
django-authtools==1.4.0

# Ratings and Reviews
django-star-ratings==0.5.1

# Slugs
django-autoslug==1.9.3
unicode-slugify==0.1.3

# Images
sorl-thumbnail==12.3